--- src/stop_start/stop_start.c
@@ -1,2 +1,0 @@
-#include "stop_start.h"
-
