--- src/stop_start/stop_start.h
@@ -1,1 +1,0 @@
-#include "../sensors/simu_sensors.h"
